% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug.R
\name{plug_list_credentials}
\alias{plug_list_credentials}
\title{List registered credentials for Plug API}
\usage{
plug_list_credentials()
}
\value{
A named list with \code{username} and \code{password} fields if credentials are found,
or an empty list if no credentials are stored.
}
\description{
This function lists all globally stored credentials (username and password) for the Plug API.
If none are found or an error occurs, it displays a message in English ("No credentials found for Plug API.")
and returns an empty list.
}
\details{
Note that the password is returned in plain text.
}
\examples{
\dontrun{
plug_list_credentials()
}
}
