% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug.R
\name{plug_store_credentials}
\alias{plug_store_credentials}
\title{Store user credentials securely for Plug API}
\usage{
plug_store_credentials(username, password)
}
\arguments{
\item{username}{The username for the Plug.}

\item{password}{The password for the Plug.}
}
\value{
No return value. The credentials are securely stored.
}
\description{
This function securely stores the global username and password required to authenticate with the Plug application.
}
\examples{
plug_store_credentials("myusername", "mypassword")
}
